spike(qwen3): Phase 3 export fp16 output + 数据 — backend ANE 9.2x 加速
改 export 脚本 outputs dtype fp32 → fp16, 重新 export mlpackage.

实测 (M1 Max, macOS 26.5, fp16 output):
- N=1 60s podcast: RTF 0.073 (跟 fp32 output 一致)
- N=1 44min 4spk: wall 343.9s, backend mlpackage 6.57s,
  vs Phase 2 N=2 同音频 60.7s ONNX backend = 9.2x 加速 ✅
- N=2 1spk-16min 跑通: wall 133s, backend 2.58s
- N=2 4spk-44min 仍 ggml_abort (跟 fp16/fp32 output 无关)

确认 ggml_abort 是 N=2 并发问题 (N=1 任何长度都 work). 推测:
N=2 时 ANE 同时跑 2 个 worker 的 frontend + backend = 4 路推理共享,
ANE 调度/数值精度在边界条件下导致 LLM sampler 越界产 invalid token id.

via [HAPI](https://hapi.run)

+++ spikes/qwen3_mac_hw_accel/phase3_backend/export_backend_coreml.py
@@ -197,7 +197,7 @@ def convert_to_mlpackage(traced, time_dim: int, dim: int, precision: str, units:
             ct.TensorType(name="hidden_states", shape=(1, time_dim, dim), dtype=np.float32),
             ct.TensorType(name="key_padding_mask", shape=(1, time_dim), dtype=np.int32),
         ],
-        outputs=[ct.TensorType(name="last_hidden_state", dtype=np.float32)],
+        outputs=[ct.TensorType(name="last_hidden_state", dtype=np.float16)],
         minimum_deployment_target=ct.target.macOS15,
         compute_units=units_map[units],
         compute_precision=precision_map[precision],
